requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
logging==0.4.9.6
minio==7.2.15
numpy==2.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
